requirements.txt
psycopg2
Django
djangorestframework